+++ src/Modules/Middleware.js
@@ -1,5 +1,0 @@
-
-const jwt = require("jsonwebtoken")
-
-function  AuthenticateToken(req, res, next){
-}