+++ src/index.js
@@ -1,4 +1,5 @@
-require("dotenv").config()
+const dotenv = require("dotenv")
+dotenv.config()
 const Account = require("./Modules/Account")
 
 const express = require("express")
@@ -9,13 +10,27 @@ const app = express()
 app.use(express.urlencoded({extended: true}))
 app.use(express.json())
 
+const index = require("./routes/index")
 const login = require("./routes/login")
 const signup = require("./routes/signup")
 const verification = require("./routes/verification")
 
+app.use("/", index)
 app.use("/login", login)
 app.use("/signup", signup)
 app.use("/verification", verification)
+
+//404 error
+app.use((req, res, next) => {
+    let ip = req.headers["x-forwarded-for"] || req.socket.remoteAddress
+    console.log(ip)
+    return res.status(404).send("URL not found")
+})
+
+
+/*
+hi, please put the send email code in ./Modules/EmailService, thanks
+*/
 
 function EnsureTokenExists(req, res, next){
     let BearerHeader = req.headers["authorization"]
@@ -31,10 +46,6 @@ function EnsureTokenExists(req, res, next){
         next()
     })
 }
-
-app.post("/api/v1/profile", EnsureTokenExists, async (req, res) => {
-
-})
 
 app.listen(process.env.PORT, () => {
     console.log(`Listening on Port ${process.env.PORT}`)